+++ packages/backend/src/api/resources/JSON.ts
@@ -2,7 +2,6 @@ import type {
   InvitationStatus,
   OrganizationInvitationStatus,
   OrganizationMembershipRole,
-  PermissionType,
   SignInStatus,
   SignUpAttributeRequirements,
   SignUpStatus,
@@ -21,8 +20,6 @@ export enum ObjectType {
   Organization = 'organization',
   OrganizationInvitation = 'organization_invitation',
   OrganizationMembership = 'organization_membership',
-  Role = 'role',
-  Permission = 'permission',
   PhoneNumber = 'phone_number',
   RedirectUrl = 'redirect_url',
   Session = 'session',
@@ -175,27 +172,6 @@ export interface OrganizationMembershipPublicUserDataJSON {
   user_id: string;
 }
 
-export interface RoleJSON extends ClerkResourceJSON {
-  object: ObjectType.Role;
-  name: string;
-  key: string;
-  description: string;
-  permissions: PermissionJSON[];
-  created_at: number;
-  updated_at: number;
-}
-
-export interface PermissionJSON extends ClerkResourceJSON {
-  object: ObjectType.Permission;
-  id: string;
-  name: string;
-  key: string;
-  description: string;
-  type: PermissionType;
-  created_at: number;
-  updated_at: number;
-}
-
 export interface PhoneNumberJSON extends ClerkResourceJSON {
   object: ObjectType.PhoneNumber;
   phone_number: string;
@@ -295,6 +271,7 @@ export interface UserJSON extends ClerkResourceJSON {
   unsafe_metadata: UserUnsafeMetadata;
   created_at: number;
   updated_at: number;
+  last_active_at: number | null;
 }
 
 export interface VerificationJSON extends ClerkResourceJSON {