+++ packages/clerk-js/src/core/clerk.ts
@@ -915,13 +915,9 @@ export class Clerk implements ClerkInterface {
         buildURL({ base: displayConfig.signInUrl, hashPath: '/reset-password' }, { stringify: true }),
     );
 
-    const navigateAfterSignIn = makeNavigate(
-      params.afterSignInUrl || params.redirectUrl || displayConfig.afterSignInUrl,
-    );
+    const navigateAfterSignIn = makeNavigate(params.afterSignInUrl || params.redirectUrl || '/');
 
-    const navigateAfterSignUp = makeNavigate(
-      params.afterSignUpUrl || params.redirectUrl || displayConfig.afterSignUpUrl,
-    );
+    const navigateAfterSignUp = makeNavigate(params.afterSignUpUrl || params.redirectUrl || '/');
 
     const navigateToContinueSignUp = makeNavigate(
       params.continueSignUpUrl ||