+++ packages/clerk-js/src/ui.retheme/styledSystem/emotion.d.ts
@@ -1,8 +1,0 @@
-// eslint-disable-next-line no-restricted-imports
-import '@emotion/react';
-
-import type { InternalTheme } from '../foundations';
-
-declare module '@emotion/react' {
-  export interface Theme extends InternalTheme {}
-}