+++ playground/chrome-extension/src/pages/devtools/index.ts
@@ -1,0 +1,6 @@
+import browser from 'webextension-polyfill';
+
+browser
+  .devtools
+  .panels
+  .create('Clerk Starter', 'icon-32.png', 'src/pages/panel/index.html');