+++ playground/chrome-extension/src/pages/panel/Panel.tsx
@@ -1,0 +1,12 @@
+import { MemoryRouter } from "react-router-dom";
+import { SharedApp } from '@/components/SharedApp';
+
+function Panel() {
+  return (
+    <MemoryRouter>
+      <SharedApp />
+    </MemoryRouter>
+  )
+}
+
+export default Panel;